Add WebTabs UI which is similar to LemonTabs but has a changable title
